chore: update Applications folder structure
