+++ src/components/Application/SystemSettings/sections/index.ts
@@ -1,0 +1,2 @@
+export { default as AppleAccount } from './AppleAccount'
+export { default as General } from './General'
